chore(tests): silence every Mojo warning across tests, examples, and lib
CI surfaced 45 warnings on the test job (no failures, but the kind of
noise that hides real regressions next week). Audited everything and
brought the warning count to zero, locally verified with the full
``pixi run tests`` + ``pixi run -e dev docs-build`` matrix (961 tests
pass, mojodoc clean).

Categorised + fixed:

- ``tests/test_server_serve_comptime.mojo``: ``if False:`` was being
  used to keep a call to ``serve_comptime[handler, config]`` alive so
  the comptime asserts would fire at build time. The compiler now
  warns on ``if False:`` as unreachable. Replaced with a runtime
  sentinel (``if port < 0:``) the compiler cannot constant-fold, so
  monomorphisation still happens but the dead-branch warning goes
  away.

- ``tests/test_header_view.mojo``,
  ``tests/test_request_view.mojo``,
  ``tests/test_request_view_reactor.mojo``: ``bytes.append(ord("X"))``
  and ``bytes_list[i] = ord("Y")`` relied on the deprecated implicit
  ``Int -> UInt8`` conversion. Wrapped each call site explicitly:
  ``bytes.append(UInt8(ord("X")))``.

- ``tests/test_request_view.mojo``: ``len(raw)`` on a ``String`` is
  discouraged; switched to ``raw.byte_length()`` (matches the rest of
  the codebase).

- ``flare/http2/frame.mojo``: ``alias H2_PREFACE = ...`` and the two
  frame-size constants were emitting "alias is deprecated; use
  comptime" through every include path. Promoted to typed
  ``comptime`` declarations (``StaticString`` / ``Int``).

- ``flare/runtime/handoff.mojo``: ``alias _MUTEX_BYTES = 64`` →
  ``comptime _MUTEX_BYTES: Int = 64``. Same warning class.

- ``flare/runtime/handoff.mojo``, ``tests/test_handoff.mojo``,
  ``examples/36_work_stealing.mojo``: ``from os import getenv/setenv``
  was emitting "implicit standard library imports are deprecated;
  fully qualify with std.". Switched to ``from std.os import ...``.

- ``tests/test_handoff.mojo``, ``examples/36_work_stealing.mojo``:
  ``setenv(...)`` returns ``Bool``; the discarded result triggers
  "Bool value is unused". Bound to ``_`` to match every other
  ``setenv`` site in the tree.

- ``tests/test_h2_server.mojo``: ``var off = 0`` was assigned and
  immediately overwritten ("assignment to 'off' was never used");
  collapsed into a single ``var off = 9 + settings.header.length``.

- ``tests/test_static_response.mojo``: docstring summary started with
  a lowercase identifier (``close()``); rewrote so it begins with
  ``\`\`close()\`\``` and the linter's "begin with capital letter or
  non-alpha character" rule is satisfied.

- ``examples/08_http_server.mojo``: the ``Response.__init__`` move
  refactor from the previous commit forced ``body`` to be either
  moved (``^``) or copied; the echo handler still needs ``req.body``
  preserved on the request object, so it now uses
  ``body=req.body.copy()``. (One-allocation copy that is intrinsic to
  the echo semantics; no perf regression on the hot path because the
  ``ok`` / ``ok_json`` helpers route through ``_string_to_bytes``.)

Verified:

- ``pixi run tests``: 961 passed, 0 failed, 0 warnings.
- ``pixi run -e dev docs-build``: built ``target/doc`` clean.

Made-with: Cursor

+++ tests/test_header_view.mojo
@@ -140,10 +140,10 @@ def test_high_bit_in_header_name_raises() raises:
     """High-bit byte in header name is non-token; reject."""
     var bytes = List[UInt8]()
     bytes.append(UInt8(0x80))
-    bytes.append(ord("X"))
-    bytes.append(ord(":"))
-    bytes.append(ord(" "))
-    bytes.append(ord("v"))
+    bytes.append(UInt8(ord("X")))
+    bytes.append(UInt8(ord(":")))
+    bytes.append(UInt8(ord(" ")))
+    bytes.append(UInt8(ord("v")))
     for b in "\r\n\r\n".as_bytes():
         bytes.append(b)
     with assert_raises():
@@ -157,8 +157,8 @@ def test_bare_cr_in_header_value_raises() raises:
     for b in "X-A: hello".as_bytes():
         bytes.append(b)
     bytes.append(UInt8(0x0D))
-    bytes.append(ord("X"))
-    bytes.append(ord("\n"))
+    bytes.append(UInt8(ord("X")))
+    bytes.append(UInt8(ord("\n")))
     for b in "\r\n".as_bytes():
         bytes.append(b)
     with assert_raises():
@@ -172,7 +172,7 @@ def test_nul_in_header_value_raises() raises:
     for b in "X-A: hello".as_bytes():
         bytes.append(b)
     bytes.append(UInt8(0))
-    bytes.append(ord("X"))
+    bytes.append(UInt8(ord("X")))
     for b in "\r\n\r\n".as_bytes():
         bytes.append(b)
     with assert_raises():
@@ -205,7 +205,7 @@ def test_into_owned_independent_lifetime() raises:
     var view = parse_header_view(Span[UInt8, _](bytes_list))
     var owned = view.into_owned()
     # Mutate the source buffer; the owned map should still see "x".
-    bytes_list[6] = ord("Y")
+    bytes_list[6] = UInt8(ord("Y"))
     assert_equal(owned.get("Host"), "x")
